// ui/src/data/decoders/generated/v2/ocap_v2.ts
export enum Side {
  SIDE_UNKNOWN = 0,
  SIDE_WEST = 1,
  SIDE_EAST = 2,
  SIDE_GUER = 3,
  SIDE_CIV = 4,
  SIDE_GLOBAL = 5,
  UNRECOGNIZED = -1,
}
// ...
export interface Position3D {
  x: number;
  y: number;
  z: number;
}

export interface Manifest {
  /** = 2 */
  version: number;
  world: WorldMeta | undefined;
  mission: MissionMeta | undefined;
  frameCount: number;
  chunkSize: number;
  captureDelayMs: number;
  chunkCount: number;
  soldiers: SoldierDef[];
  vehicles: VehicleDef[];
  events: Event[];
  markers: MarkerDef[];
  times: TimeSample[];
}

export interface WorldMeta {
  worldName: string;
  worldSize: number;
  latitude: number;
  longitude: number;
  author: string;
  displayName: string;
}

export interface MissionMeta {
  missionName: string;
  briefingName: string;
  author: string;
  serverName: string;
  extensionVersion: string;
  addonVersion: string;
  extensionBuild: string;
  tag: string;
  playableSlots: PlayableSlots | undefined;
  sideFriendly: SideFriendly | undefined;
  addons: Addon[];
}

export interface PlayableSlots {
  west: number;
  east: number;
  independent: number;
  civilian: number;
}

export interface SideFriendly {
  eastWest: boolean;
  eastIndependent: boolean;
  westIndependent: boolean;
}

export interface Addon {
  name: string;
  workshopId: string;
}

export interface SoldierDef {
  id: number;
  name: string;
  side: Side;
  groupName: string;
  role: string;
  startFrame: number;
  endFrame: number;
  isPlayer: boolean;
  className: string;
  playerUid: string;
  framesFired: FiredFrame[];
}

export interface VehicleDef {
  id: number;
  name: string;
  /** ocapType: car, tank, heli, etc. */
  vehicleClass: string;
  className: string;
  startFrame: number;
  endFrame: number;
  customization: string;
}

export interface FiredFrame {
  frameNum: number;
  startPos: Position3D | undefined;
  endPos: Position3D | undefined;
  weapon: string;
  magazine: string;
  firingMode: string;
}

export interface Event {
  frameNum: number;
  kill?: KillEvent | undefined;
  hit?: HitEvent | undefined;
  projectile?: ProjectileEvent | undefined;
  chat?: ChatEvent | undefined;
  radio?: RadioEvent | undefined;
  ace3Death?: Ace3DeathEvent | undefined;
  ace3Unconscious?: Ace3UnconsciousEvent | undefined;
  general?: GeneralEvent | undefined;
  connect?: ConnectEvent | undefined;
  endMission?: EndMissionEvent | undefined;
  telemetry?: TelemetryEvent | undefined;
}
// ...
export interface Chunk {
  index: number;
  startFrame: number;
  frameCount: number;
  frames: Frame[];
}

export interface Frame {
  frameNum: number;
  soldiers: SoldierState[];
  vehicles: VehicleState[];
}

export interface SoldierState {
  id: number;
  position: Position3D | undefined;
  bearing: number;
  /** 0=dead, 1=alive, 2=unconscious */
  lifestate: number;
  vehicleId: number;
  inVehicle: boolean;
  name: string;
  isPlayer: boolean;
  groupName: string;
  side: string;
  role: string;
  vehicleRole: string;
  stance: string;
  hasStableVitals: boolean;
  isDraggedCarried: boolean;
  scores: SoldierScores | undefined;
}

export interface SoldierScores {
  infantryKills: number;
  vehicleKills: number;
  armorKills: number;
  airKills: number;
  deaths: number;
  totalScore: number;
}

export interface VehicleState {
  id: number;
  position: Position3D | undefined;
  bearing: number;
  alive: boolean;
  crewIds: number[];
  fuel: number;
  damage: number;
  locked: boolean;
  engineOn: boolean;
  side: string;
  turretAzimuth: number;
  turretElevation: number;
}

export interface MarkerDef {
  type: string;
  text: string;
  startFrame: number;
  endFrame: number;
  playerId: number;
  color: string;
  side: Side;
  positions: MarkerPosition[];
  size: number[];
  shape: string;
  brush: string;
}

export interface MarkerPosition {
  frameNum: number;
  position: Position3D | undefined;
  direction: number;
  alpha: number;
  lineCoords: number[];
}

export interface TimeSample {
  frameNum: number;
  systemTimeUtc: string;
  date: string;
  timeMultiplier: number;
  time: number;
}
// ...
export const VehicleDef: MessageFns<VehicleDef> = {
  encode(message: VehicleDef, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== 0) {
      writer.uint32(8).uint32(message.id);
    }
    if (message.name !== "") {
      writer.uint32(18).string(message.name);
    }
    if (message.vehicleClass !== "") {
      writer.uint32(26).string(message.vehicleClass);
    }
    if (message.className !== "") {
      writer.uint32(34).string(message.className);
    }
    if (message.startFrame !== 0) {
      writer.uint32(40).uint32(message.startFrame);
    }
    if (message.endFrame !== 0) {
      writer.uint32(48).uint32(message.endFrame);
    }
    if (message.customization !== "") {
      writer.uint32(58).string(message.customization);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): VehicleDef {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseVehicleDef();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.id = reader.uint32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.vehicleClass = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.className = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.startFrame = reader.uint32();
          continue;
        }
        case 6: {
          if (tag !== 48) {
            break;
          }

          message.endFrame = reader.uint32();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.customization = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): VehicleDef {
    return {
      id: isSet(object.id) ? globalThis.Number(object.id) : 0,
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      vehicleClass: isSet(object.vehicleClass)
        ? globalThis.String(object.vehicleClass)
        : isSet(object.vehicle_class)
        ? globalThis.String(object.vehicle_class)
        : "",
      className: isSet(object.className)
        ? globalThis.String(object.className)
        : isSet(object.class_name)
        ? globalThis.String(object.class_name)
        : "",
      startFrame: isSet(object.startFrame)
        ? globalThis.Number(object.startFrame)
        : isSet(object.start_frame)
        ? globalThis.Number(object.start_frame)
        : 0,
      endFrame: isSet(object.endFrame)
        ? globalThis.Number(object.endFrame)
        : isSet(object.end_frame)
        ? globalThis.Number(object.end_frame)
        : 0,
      customization: isSet(object.customization) ? globalThis.String(object.customization) : "",
    };
  },

  toJSON(message: VehicleDef): unknown {
    const obj: any = {};
    if (message.id !== 0) {
      obj.id = Math.round(message.id);
    }
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.vehicleClass !== "") {
      obj.vehicleClass = message.vehicleClass;
    }
    if (message.className !== "") {
      obj.className = message.className;
    }
    if (message.startFrame !== 0) {
      obj.startFrame = Math.round(message.startFrame);
    }
    if (message.endFrame !== 0) {
      obj.endFrame = Math.round(message.endFrame);
    }
    if (message.customization !== "") {
      obj.customization = message.customization;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<VehicleDef>, I>>(base?: I): VehicleDef {
    return VehicleDef.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<VehicleDef>, I>>(object: I): VehicleDef {
    const message = createBaseVehicleDef();
    message.id = object.id ?? 0;
    message.name = object.name ?? "";
    message.vehicleClass = object.vehicleClass ?? "";
    message.className = object.className ?? "";
    message.startFrame = object.startFrame ?? 0;
    message.endFrame = object.endFrame ?? 0;
    message.customization = object.customization ?? "";
    return message;
  },
};
// ...
export const Chunk: MessageFns<Chunk> = {
  encode(message: Chunk, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.index !== 0) {
      writer.uint32(8).uint32(message.index);
    }
    if (message.startFrame !== 0) {
      writer.uint32(16).uint32(message.startFrame);
    }
    if (message.frameCount !== 0) {
      writer.uint32(24).uint32(message.frameCount);
    }
    for (const v of message.frames) {
      Frame.encode(v!, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Chunk {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseChunk();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.index = reader.uint32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.startFrame = reader.uint32();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.frameCount = reader.uint32();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.frames.push(Frame.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Chunk {
    return {
      index: isSet(object.index) ? globalThis.Number(object.index) : 0,
      startFrame: isSet(object.startFrame)
        ? globalThis.Number(object.startFrame)
        : isSet(object.start_frame)
        ? globalThis.Number(object.start_frame)
        : 0,
      frameCount: isSet(object.frameCount)
        ? globalThis.Number(object.frameCount)
        : isSet(object.frame_count)
        ? globalThis.Number(object.frame_count)
        : 0,
      frames: globalThis.Array.isArray(object?.frames) ? object.frames.map((e: any) => Frame.fromJSON(e)) : [],
    };
  },

  toJSON(message: Chunk): unknown {
    const obj: any = {};
    if (message.index !== 0) {
      obj.index = Math.round(message.index);
    }
    if (message.startFrame !== 0) {
      obj.startFrame = Math.round(message.startFrame);
    }
    if (message.frameCount !== 0) {
      obj.frameCount = Math.round(message.frameCount);
    }
    if (message.frames?.length) {
      obj.frames = message.frames.map((e) => Frame.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Chunk>, I>>(base?: I): Chunk {
    return Chunk.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Chunk>, I>>(object: I): Chunk {
    const message = createBaseChunk();
    message.index = object.index ?? 0;
    message.startFrame = object.startFrame ?? 0;
    message.frameCount = object.frameCount ?? 0;
    message.frames = object.frames?.map((e) => Frame.fromPartial(e)) || [];
    return message;
  },
};
// ...
export const MarkerDef: MessageFns<MarkerDef> = {
// ...
  fromJSON(object: any): MarkerDef {
    return {
// ...
      shape: isSet(object.shape) ? globalThis.String(object.shape) : "",
      brush: isSet(object.brush) ? globalThis.String(object.brush) : "",
    };
  },
// ...
};
// ...
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}